responsive title
