+++ package.json
@@ -12,6 +12,7 @@
   "dependencies": {
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-hook-form": "^7.50.0",
     "react-icons": "^5.0.1",
     "react-router-dom": "^6.21.3"
   },
