+++ src/components/pages/Contact.jsx
@@ -1,5 +1,19 @@
+/**
+ * The external imports
+ */
+
+
+/**
+ * The internal imports
+ */
+import MyTitle from "../MyTitle";
+
 export default function Contact() {
   return (
-    <div>Contact</div>
-  )
+    <div className="w-[90%] m-auto">
+      <MyTitle span1={"prendre"} span2={"contact"} shadow={"contact"} />
+      <form action=""></form>
+      
+    </div>
+  );
 }
