+++ src/styles/fonts.css
@@ -21,7 +21,6 @@ h1, h2, h3, h4, h5, h6 {
 h1 {
   line-height: 3.444rem;
   font-size: 2.833rem;
-  letter-spacing: -1px;
 }
 
 h2 {
